% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_ops.R
\name{ask_for_key}
\alias{ask_for_key}
\title{Ask for passkey}
\usage{
ask_for_key()
}
\value{
(cyphr_key) sodium key for encrypting data with \link[cyphr:encrypt]{cyphr::encrypt}
}
\description{
\code{ask_for_key} asks for a passkey which will be used to create a sodium
symmetric key (using \link[cyphr:key_sodium]{cyphr::key_sodium}) which can then be used to encrypt
your data using \link[cyphr:encrypt]{cyphr::encrypt}.
}
\examples{
try(
  if (!exists("key")) {
    key <- ask_for_key()
    print(key)
  }
)

}
