% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{do_multicollinearity_test}
\alias{do_multicollinearity_test}
\title{Perform multicollinearity test on a generalised linear model}
\usage{
do_multicollinearity_test(model, vif_threshold = 10)
}
\arguments{
\item{model}{(glm object) A \link[stats:glm]{stats::glm} model object}

\item{vif_threshold}{(numeric) The threshold above which is suggestive of the
presence of multicollinearity. Typical values are 5 or 10. The default value is 10.}
}
\value{
None
}
\description{
\code{do_multicollinearity_test} is a wrapper function that prints the results from
\link[car:vif]{car::vif} and \link[mctest:imcdiag]{mctest::imcdiag}
}
\examples{
if (require("mlbench")) {
  # Load data -----------------------------------------------------------------
  data("PimaIndiansDiabetes2", package = "mlbench")

  # Fit a logistic regression model -------------------------------------------
  model <- glm(diabetes ~ age + pregnant + glucose + pressure + mass + pedigree,
    data = PimaIndiansDiabetes2,
    family = binomial(link = "logit")
  )

  # Run the test for multicollinearity ----------------------------------------
  model \%>\% do_multicollinearity_test()
}

}
