% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_ops.R
\name{export_excel_to_rds}
\alias{export_excel_to_rds}
\title{Export excel sheets to RDS files}
\usage{
export_excel_to_rds(filepath, encrypt_file = TRUE, key = NULL, ...)
}
\arguments{
\item{filepath}{(character) File path to excel workbook}

\item{encrypt_file}{(logical) If TRUE, each RDS file will be encrypted. See
\link{ezyr_save_rds} which will be called internally}

\item{key}{(character) Optional argument for \link{ask_for_key}, which will be
called internally. Defaults to NULL, for which the user will be prompted for
a passkey at runtime via \link[rstudioapi:askForPassword]{rstudioapi::askForPassword}}

\item{...}{(ellipsis) Extra arguments to pass into \link[readxl:read_excel]{readxl::read_excel},
which will be called internally}
}
\value{
None
}
\description{
\code{export_excel_to_rds} exports each excel sheet in an excel workbook into
a separate RDS file.
}
\examples{
library(mlbench)
data("BostonHousing2")
data("Glass")
data("Ozone")

# Create excel workbook with 3 sheets of data -------------------------------
excel_workbook_filepath <- file.path(tempdir(), "myfile.xlsx")
export_to_excel(BostonHousing2, excel_workbook_filepath, sheet_name = "BH2")
export_to_excel(Glass, excel_workbook_filepath, sheet_name = "Glass")
export_to_excel(Ozone, excel_workbook_filepath, sheet_name = "Ozone")

# Run export_excel_to_rds ---------------------------------------------------
## You may opt to omit "key" argument and get prompted for it at runtime
export_excel_to_rds(excel_workbook_filepath, encrypt = TRUE, key = "12345")
print(list.files(tempdir(), pattern = "\\\\.(xlsx|xls|rds)$"))

# Delete the created files in temp directory --------------------------------
unlink(excel_workbook_filepath)
unlink(file.path(tempdir(), "myfile_bh2.rds"))
unlink(file.path(tempdir(), "myfile_glass.rds"))
unlink(file.path(tempdir(), "myfile_ozone.rds"))
print(list.files(tempdir(), pattern = "\\\\.(xlsx|xls|rds)$"))

}
