% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_ops.R
\name{export_to_excel}
\alias{export_to_excel}
\title{Export to excel}
\usage{
export_to_excel(
  output_table,
  workbook,
  sheet_name = format(Sys.time(), "\%d\%b\%Y_\%H\%M")
)
}
\arguments{
\item{output_table}{(data.frame) Data that you want to export to excel}

\item{workbook}{(character) File path of excel workbook
\itemize{
\item If the workbook does not exist, it will first be created
\item If the workbook exists, data will be exported to the specified sheet
}}

\item{sheet_name}{(character) Name of excel sheet to export data to
\itemize{
\item Defaults to system time at execution if none provided
}}
}
\value{
None
}
\description{
\code{export_to_excel} enables you to export your data out to an excel workbook.
}
\examples{
# Load data
data(BostonHousing2, package = "mlbench")

# Export data to an excel file in temp directory
output_filepath_1 <- file.path(tempdir(), "file1.xlsx")
export_to_excel(BostonHousing2, output_filepath_1, sheet_name = "my_sheet")
print(list.files(tempdir(), pattern = "\\\\.(xlsx|xls)$"))

# Delete the excel file in temp directory
unlink(output_filepath_1)
print(list.files(tempdir(), pattern = "\\\\.(xlsx|xls)$"))

}
