% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ops.R
\name{extend_paired}
\alias{extend_paired}
\title{Extend a paired table}
\usage{
extend_paired(paired_obj, labels = NULL, bold_var = TRUE)
}
\arguments{
\item{paired_obj}{(paired object) An \link[arsenal:paired]{arsenal::paired} object}

\item{labels}{(list) An optional named list of variable labels, for the data
passed into \code{paired_obj}. \code{labels} may be set with \link[labelled:var_label]{labelled::var_label}}

\item{bold_var}{(logical) If TRUE, variable sub-headers will be wrapped with **}
}
\value{
(data.frame) A paired table appended with marker
}
\description{
\code{extend_paired} is a wrapper around \link[arsenal:paired]{arsenal::paired} that adds a marker
which enables the joining of two paired tables together.
}
\examples{
library(arsenal)
library(tidyverse)
library(labelled)

# Generate fake patient data for illustration -------------------------------
generate_fake_patient_data <- function(id) {
  data.frame(list(
    patient_id = id,
    time = c("baseline", "12-month"),
    hba1c = runif(n = 2, min = 4, max = 14),
    ldl = sample(c("good control", "bad control"))
  ))
}

set.seed(2024)
data <- map(
  sprintf("P\%s", str_pad(1:30, width = 2, pad = 0)),
  generate_fake_patient_data
) \%>\%
  bind_rows()
print(head(data, 10))

# Create paired table without labels ----------------------------------------
paired(time ~ hba1c + ldl,
  data = data \%>\% mutate(across(
    .cols = time,
    .fns = ~ fct_relevel(.x, "baseline", "12-month")
  )),
  id = patient_id,
  control = paired.control(
    digits = 2,
    numeric.stats = c("Nmiss", "meansd", "medianq1q3", "range"),
    numeric.test = "signed.rank",
    signed.rank.correct = FALSE
  )
) \%>\%
  extend_paired()

# Create paired table with labels -------------------------------------------
## Add labels
var_label(data) <- list(
  hba1c = "Hemoglobin A1c",
  ldl = "LDL control"
)

## Create paired table
paired(time ~ hba1c + ldl,
  data = data \%>\% mutate(across(
    .cols = time,
    .fns = ~ fct_relevel(.x, "baseline", "12-month")
  )),
  id = patient_id,
  control = paired.control(
    digits = 2,
    numeric.stats = c("Nmiss", "meansd", "medianq1q3", "range"),
    numeric.test = "signed.rank",
    signed.rank.correct = FALSE
  )
) \%>\%
  extend_paired(labels = var_label(data))

}
