% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ops.R
\name{extend_tableby}
\alias{extend_tableby}
\title{Extend the output of the tableby function}
\usage{
extend_tableby(tableby_obj, smd = TRUE, smd_digits = 2, bold_var = TRUE)
}
\arguments{
\item{tableby_obj}{(tableby object) An \link[arsenal:tableby]{arsenal::tableby} object}

\item{smd}{(logical) If TRUE, computes the standardised mean difference (SMD)}

\item{smd_digits}{(numeric) Number of decimal places to express SMD}

\item{bold_var}{(logical) If TRUE, variables will be wrapped with **}
}
\value{
(data.frame) A tableby table appended with marker and SMD
}
\description{
\code{extend_tableby} is a wrapper around \link[arsenal:tableby]{arsenal::tableby} that adds the
following to a tableby table:
\enumerate{
\item Marker for enabling the joining of two tableby tables together
\item Standardised mean difference (SMD) between the groups
}
}
\examples{
library(arsenal)
library(MatchIt)
data(lalonde)

tableby(treat ~ age + educ + race + married,
  data = lalonde,
  control = tableby.control(numeric.stats = c("Nmiss", "meansd", "medianq1q3", "range"))
) \%>\%
  extend_tableby()

}
