% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ops.R
\name{extract_reading_from_interval}
\alias{extract_reading_from_interval}
\title{Extract latest reading in specified window period of observation}
\usage{
extract_reading_from_interval(
  df_master,
  df_master_fields,
  df_long,
  df_long_fields,
  params
)
}
\arguments{
\item{df_master}{(data.frame) Master dataframe to join the readings unto.
\emph{df_master} must minimally contain the following 2 columns:
\itemize{
\item A column of IDs to identify each patient
\item A column of index dates (represented by T as described above) for each patient
}}

\item{df_master_fields}{(list) A named list to identify the \emph{ID} and
\emph{index date} columns in \emph{df_master}
\itemize{
\item It takes on this format: \verb{list(ID = <your_ID_column>, index_date = <your_index_date_column>)}
\item For example: \code{list(ID = "patient_id", index_date = "enrolment_date")}
}}

\item{df_long}{(data.frame) Longitudinal data e.g. laboratory readings}

\item{df_long_fields}{(list) A named list to identify the \emph{reading date} and \emph{reading(s)} in \emph{df_long}
\itemize{
\item It takes on this format: \verb{list(reading_date = <your_reading_date_column>, reading = <your_reading_column(s)>)}
\itemize{
\item \emph{reading_date} (character) The name of the dataframe column that contains the dates of the readings
\item \emph{reading} (character or character vector) The name(s) of the dataframe column(s) that contains the readings
}
\item For example: \code{list(reading_date = "calendar_date", reading = c("bp_systolic", "bp_diastolic"))}
}}

\item{params}{(list) A named list where the names are the timepoints and the
values are numeric vectors that represent \code{c(lower_bound, upper_bound)}
\itemize{
\item For example: \code{params <- list(baseline = c(-180, 0), "3mth" = c(31, 90), "6mth" = c(120, 180))}
}}
}
\value{
(data.frame) \emph{df_master} with longitudinal readings joined unto
}
\description{
Suppose that you have the following, \code{extract_reading_from_interval} will
extract the latest laboratory reading for each patient within the respective
window periods to be considered as the "baseline" reading, "3-month" reading etc.
\enumerate{
\item a set of longitudinal data e.g. laboratory readings of patients
\item a set of parameters that define the window period of observation for
considering readings that would constitute as "baseline" reading,
"3-month" reading etc. for example:
\itemize{
\item \emph{baseline}: T-180 days to T
\item \emph{3-month}: T+31 days to T+90 days
\item where T is the index date which could be the enrolment date or
date of first visit
}
}
}
\examples{
library(dplyr)

# Generate fake patient data for illustration -------------------------------
generate_fake_patient_data <- function(id) {
  data.frame(
    list(
      patient_id = id,
      enrolment_date = Sys.Date() + runif(1, min = -730, max = 0),
      bp_systolic = sample(30:160, 9, replace = FALSE),
      bp_diastolic = sample(10:100, 9, replace = FALSE),
      hba1c = sample(4:14, 9, replace = FALSE)
    ),
    time = c(rep("baseline", 3), rep("6mth", 3), rep("12mth", 3))
  ) \%>\%
    mutate(visit_date = case_when(
      time == "baseline" ~ enrolment_date + sample(-180:0, 3, replace = FALSE),
      time == "6mth" ~ enrolment_date + sample(100:180, 3, replace = FALSE),
      time == "12mth" ~ enrolment_date + sample(240:360, 3, replace = FALSE)
    ))
}

df_long <- rbind(
  generate_fake_patient_data("P01"),
  generate_fake_patient_data("P02"),
  generate_fake_patient_data("P03")
) \%>\%
  dplyr::arrange(patient_id, visit_date)

# Master dataframe
df_master <- df_long \%>\% distinct(patient_id, enrolment_date)
print(df_master)

# Longitudinal data of readings
df_long <- df_long \%>\% select(-enrolment_date)
print(df_long)

# Extract readings ----------------------------------------------------------
df_master \%>\%
  # Blood pressure
  extract_reading_from_interval(
    list(
      ID = "patient_id",
      index_date = "enrolment_date"
    ),
    df_long,
    list(
      reading_date = "visit_date",
      reading = c("bp_systolic", "bp_diastolic")
    ),
    list(
      baseline = c(-180, 0),
      "6mth" = c(100, 180),
      "12mth" = c(240, 360)
    )
  ) \%>\%
  # HbA1c
  extract_reading_from_interval(
    list(
      ID = "patient_id",
      index_date = "enrolment_date"
    ),
    df_long,
    list(
      reading_date = "visit_date",
      reading = "hba1c"
    ),
    list(
      baseline = c(-180, 0),
      "6mth" = c(100, 180),
      "12mth" = c(240, 360)
    )
  ) \%>\%
  print()

}
