% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezyr-package.R
\docType{package}
\name{ezyr-package}
\alias{ezyr}
\alias{ezyr-package}
\title{ezyr: Helper Function Toolkit That Makes Data Operations Easier}
\description{
This package contains helper functions that make it easier to handle, analyse, and present data.
}
\author{
\strong{Maintainer}: Jeremy Lew \email{first@example.com}

}
\keyword{internal}
