% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_ops.R
\name{ezyr_decrypt_rds}
\alias{ezyr_decrypt_rds}
\title{Decrypt an encrypted RDS file}
\usage{
ezyr_decrypt_rds(filepath, key = NULL)
}
\arguments{
\item{filepath}{(character) File path to RDS file}

\item{key}{(character) Optional argument for \link{ask_for_key}, which will be
called internally. Defaults to NULL, for which the user will be prompted for
a passkey at runtime via \link[rstudioapi:askForPassword]{rstudioapi::askForPassword}}
}
\value{
Decrypted RDS file
}
\description{
Decrypt an encrypted RDS file
}
\examples{
# ezyr_decrypt_rds("C:/Users/<your_username>/.../your_encrypted_file.rds")

}
