% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_p_value}
\alias{format_p_value}
\title{Format p-value}
\usage{
format_p_value(p_value)
}
\arguments{
\item{p_value}{(numeric or numeric vector) p-value}
}
\value{
(character or character vector) formatted p-value
}
\description{
\code{format_p_value} converts p-values in the raw numeric form
to a string for presenting in tables.
}
\examples{
format_p_value(0.0001)
format_p_value(0.0086)
format_p_value(0.0097)
format_p_value(0.0456)
format_p_value(0.0495)

# Rounding follows IEEE 754 standard so rounding of 5 may not be up
format_p_value(0.0085)
format_p_value(0.055)

}
