% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{get_presentation_template_mlogit}
\alias{get_presentation_template_mlogit}
\title{Get presentation template for a multinomial logistic regression model}
\usage{
get_presentation_template_mlogit(model, data_labels = NULL)
}
\arguments{
\item{model}{(mlogit object) A \link[mlogit:mlogit]{mlogit::mlogit} model object}

\item{data_labels}{(list) A named list of labels
\itemize{
\item names: variable names
\item values: labels
\item If the labels were set with the labelled package,
the list is obtained by invoking \code{\link[labelled:var_label]{labelled::var_label()}}
}}
}
\value{
(data.frame) Table of your multinomial logistic regression results
}
\description{
\code{get_presentation_template_mlogit} creates a data.frame
template that lays out the variables in your model.
}
\examples{
library(mlogit)
data(iris)

# Run a multinomial logistic regression on iris dataset ---------------------
mlogit(Species ~ 1 | Sepal.Length + Sepal.Width + Petal.Length + Petal.Width,
  data = mlogit.data(
    iris \%>\%
      mutate(Petal.Length = ifelse(
        Petal.Length < 2.5,
        "Petal length < 2.5",
        "Petal length >= 2.5"
      ) \%>\%
        forcats::fct_relevel("Petal length < 2.5", "Petal length >= 2.5")),
    choice = "Species",
    shape = "wide"
  ),
  reflevel = "setosa",
  na.action = na.omit
) \%>\%
  get_presentation_template_mlogit()

# Run a multinomial logistic regression on labelled iris dataset ------------
labels <- list(
  Sepal.Length = "Sepal length",
  Sepal.Width = "Sepal width",
  Petal.Length = "Petal length",
  Petal.Width = "Petal width"
)
labelled::var_label(iris) <- labels
mlogit(Species ~ 1 | Sepal.Length + Sepal.Width + Petal.Length + Petal.Width,
  data = mlogit.data(
    iris \%>\%
      mutate(Petal.Length = ifelse(
        Petal.Length < 2.5,
        "Petal length < 2.5",
        "Petal length >= 2.5"
      ) \%>\%
        forcats::fct_relevel("Petal length < 2.5", "Petal length >= 2.5")),
    choice = "Species",
    shape = "wide"
  ),
  reflevel = "setosa",
  na.action = na.omit
) \%>\%
  get_presentation_template_mlogit(data_labels = labelled::var_label(iris))

}
