% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{get_regression_presentation_template}
\alias{get_regression_presentation_template}
\title{Get presentation template for a generalised linear model}
\usage{
get_regression_presentation_template(model)
}
\arguments{
\item{model}{(glm object) A \link[stats:glm]{stats::glm} model object}
}
\value{
(data.frame) A template for you to left join unto to present your regression results
}
\description{
\code{get_regression_presentation_template} creates a data.frame template that
lays out the variables in your model e.g. levels of a factor variable.
}
\details{
\code{get_regression_presentation_template} has been incorporated into \link{tabulate_glm_result},
to create a table of your GLM results for presentation in an Rmarkdown document.
}
\examples{
library(dplyr)
if (require("mlbench") & require("forcats")) {
  data(BostonHousing2, package = "mlbench")

  glm(medv ~ crim + zn + indus + town + rm + age + dis,
    data = BostonHousing2 \%>\%
      filter(town \%in\% c(
        "Newton", "Boston South Boston", "Boston Roxbury",
        "Somerville", "Boston Savin Hill", "Cambridge"
      )) \%>\%
      dplyr::mutate_at("town", forcats::fct_drop),
    family = gaussian(link = "identity")
  ) \%>\%
    get_regression_presentation_template()
}

}
