% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ops.R
\name{label_ix_columns}
\alias{label_ix_columns}
\title{Label prefixed or suffixed dataframe columns}
\usage{
label_ix_columns(df, pattern, suffix = TRUE)
}
\arguments{
\item{df}{(data.frame) Your data in the form of a data frame}

\item{pattern}{(character) regex pattern to identify your prefix/suffix}

\item{suffix}{(logical) TRUE if the pattern is a suffix.
FALSE if the pattern is a prefix}
}
\value{
(data.frame) Your data with new labels added
}
\description{
\code{label_ix_columns} adds labels to column(s) of your data.frame which have the
particular prefix/suffix you specified, by copying the labels from
corresponding columns without the prefix/suffix.
The labels are added with \code{\link[labelled:var_label]{labelled::var_label()}}.
}
\examples{
library(labelled)

# Example of suffix ---------------------------------------------------------
df <- data.frame(
  sex = sample(1:2, 5, replace = TRUE),
  race = sample(1:4, 5, replace = TRUE)
)
df$sex.factor <- c("Female", "Male")[df$sex]
df$race.factor <- c("Chinese", "Malay", "Indian", "Others")[df$race]

# Set labels
labels <- list(
  sex = "Sex",
  race = "Race"
)
var_label(df) <- labels

# Get labels after label_ix_columns()
var_label(label_ix_columns(df, ".factor", suffix = TRUE))

# Example of prefix ---------------------------------------------------------
df <- data.frame(
  sex = sample(1:2, 5, replace = TRUE),
  race = sample(1:4, 5, replace = TRUE)
)
df$fct_sex <- c("Female", "Male")[df$sex]
df$fct_race <- c("Chinese", "Malay", "Indian", "Others")[df$race]

# Set labels
labels <- list(
  sex = "Sex",
  race = "Race"
)
var_label(df) <- labels

# Get labels after label_ix_columns()
var_label(label_ix_columns(df, "fct_", suffix = FALSE))

}
