% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_propensity_score_histogram}
\alias{plot_propensity_score_histogram}
\title{Plot a histogram from a MatchIt object}
\usage{
plot_propensity_score_histogram(
  matchit_obj,
  trt_control_var,
  trt_label,
  control_label,
  is_after_matching = TRUE,
  num_bins = 30
)
}
\arguments{
\item{matchit_obj}{(matchit object) A \link[MatchIt:matchit]{MatchIt::matchit} object from the MatchIt R package}

\item{trt_control_var}{(quosure) variable that stores the treatment/control grouping}

\item{trt_label}{(character) value for treatment group in trt_control_var (e.g. video_consult_grp)}

\item{control_label}{(character) value for control group in trt_control_var (e.g. face_to_face_grp)}

\item{is_after_matching}{(logical) TRUE for after-matching plot, FALSE for before-matching plot}

\item{num_bins}{(numeric) number of bins of the histogram}
}
\value{
ggplot object
}
\description{
\code{plot_propensity_score_histogram} plots a mirrored histogram which
enables visual inspection of the distribution of propensity scores
in the treatment vs control group.
}
\examples{
library(MatchIt)

MatchIt::matchit(
  treat ~ age + educ + race + married + nodegree + re78,
  data = lalonde \%>\% dplyr::mutate(trt_control = ifelse(treat == 1, "trt", "control")),
  method = "nearest",
  ratio = 2
) \%>\%
  plot_propensity_score_histogram(trt_control, "trt", "control",
    is_after_matching = TRUE, num_bins = 30
  )
}
