% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ops.R
\name{rename_colnames}
\alias{rename_colnames}
\title{Rename column names}
\usage{
rename_colnames(df, separator = "_", add_labels = TRUE)
}
\arguments{
\item{df}{(data.frame) Your data in the form of a data frame}

\item{separator}{(character) Separator to replace whitespaces}

\item{add_labels}{(logical) TRUE if you want original
column names to be used as labels, FALSE otherwise}
}
\value{
(data.frame) Your data with cleaned column names
}
\description{
\code{rename_colnames} cleans up column names of your raw data frame with
\link{make.names}, replaces whitespaces with a separator of your choice
and converts to lower case.
}
\examples{
if (require("tibble")) {
  print("Original data")
  data <- tibble::tibble(
    "Column A!!!" = sample(1:100, 5, replace = TRUE),
    "Column - 2;" = runif(5, min = 0, max = 1),
    "Column: (3)" = sample(letters, 5, replace = TRUE),
    "<Column 4...>" = sample(c(TRUE, FALSE), 5, replace = TRUE)
  )
  data \%>\% print()

  print("Data after renaming columns")
  data \%>\%
    rename_colnames() \%>\%
    print()
}

}
