% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{tabulate_mlogit_result}
\alias{tabulate_mlogit_result}
\title{Tabulate multinomial logistic regression result}
\usage{
tabulate_mlogit_result(
  model,
  outcome_levels_2onwards,
  data_labels = NULL,
  num_dp = 2
)
}
\arguments{
\item{model}{(mlogit object) A \link[mlogit:mlogit]{mlogit::mlogit} model object}

\item{outcome_levels_2onwards}{(character) A character vector of outcome levels, from level 2 onwards
\enumerate{
\item outcome level 1 refers to the reference level (omitted)
\item outcome levels 2 onwards refers to other outcome levels other than the reference level
\item the results for each of the outcome levels will be tabulated in order, from left to right
}}

\item{data_labels}{(list) A named list of labels
\itemize{
\item names: variable names
\item values: labels
\item If the labels were set with the labelled package,
the list is obtained by invoking \code{\link[labelled:var_label]{labelled::var_label()}}
}}

\item{num_dp}{(numeric) Number of decimal places to present confidence intervals. Defaults to 2 d.p.}
}
\value{
(data.frame) Table of your multinomial logistic regression results
}
\description{
\code{tabulate_mlogit_result} tabulates the results of your multinomial logistic regression
that were created from running \link[mlogit:mlogit]{mlogit::mlogit}.
}
\examples{
library(mlogit)
data(iris)

# Run a multinomial logistic regression on iris dataset ---------------------
mlogit(Species ~ 1 | Sepal.Length + Sepal.Width + Petal.Length + Petal.Width,
  data = mlogit.data(
    iris \%>\%
      mutate(Petal.Length = ifelse(
        Petal.Length < 2.5,
        "Petal length < 2.5",
        "Petal length >= 2.5"
      ) \%>\%
        forcats::fct_relevel("Petal length < 2.5", "Petal length >= 2.5")),
    choice = "Species",
    shape = "wide"
  ),
  reflevel = "setosa",
  na.action = na.omit
) \%>\%
  tabulate_mlogit_result(c("versicolor", "virginica"))

# Run a multinomial logistic regression on labelled iris dataset ------------
labels <- list(
  Sepal.Length = "Sepal length",
  Sepal.Width = "Sepal width",
  Petal.Length = "Petal length",
  Petal.Width = "Petal width"
)
labelled::var_label(iris) <- labels
mlogit(Species ~ 1 | Sepal.Length + Sepal.Width + Petal.Length + Petal.Width,
  data = mlogit.data(
    iris \%>\%
      mutate(Petal.Length = ifelse(
        Petal.Length < 2.5,
        "Petal length < 2.5",
        "Petal length >= 2.5"
      ) \%>\%
        forcats::fct_relevel("Petal length < 2.5", "Petal length >= 2.5")),
    choice = "Species",
    shape = "wide"
  ),
  reflevel = "setosa",
  na.action = na.omit
) \%>\%
  tabulate_mlogit_result(c("versicolor", "virginica"), data_labels = labelled::var_label(iris))

}
