% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_ops.R
\name{write_to_excel}
\alias{write_to_excel}
\title{Write to excel}
\usage{
write_to_excel(
  output_table,
  workbook,
  sheet_name = format(Sys.time(), "\%d\%b\%Y_\%H\%M")
)
}
\arguments{
\item{output_table}{(data.frame) Data that you want to write to excel}

\item{workbook}{(character) File path of excel workbook
\itemize{
\item If the workbook does not exist, it will be created
\item If the workbook exists, data will be written to the specified sheet
}}

\item{sheet_name}{(character) Name of excel sheet to write data to
\itemize{
\item Defaults to system time at execution if none provided
}}
}
\value{
None
}
\description{
\code{write_to_excel} enables you to write your data out to an excel workbook.
}
\examples{
# Load data
data(BostonHousing2, package = "mlbench")

# Write data to excel file in temp directory
output_file_path_1 <- file.path(tempdir(), "file1.xlsx")
write_to_excel(BostonHousing2, output_file_path_1, sheet_name = "my_sheet")
print(list.files(tempdir(), pattern = "\\\\.(xlsx|xls)$"))

# Clean up file in temp directory
unlink(output_file_path_1)
print(list.files(tempdir(), pattern = "\\\\.(xlsx|xls)$"))

}
