% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ops.R
\name{extend_tableby}
\alias{extend_tableby}
\title{Extend a tableby table}
\usage{
extend_tableby(tableby_obj, data, smd = TRUE, smd_digits = 2, bold_var = TRUE)
}
\arguments{
\item{tableby_obj}{(tableby object) An \link[arsenal:tableby]{arsenal::tableby} object}

\item{data}{(data.frame) Data passed into \code{tableby_obj}}

\item{smd}{(logical) If TRUE, computes the standardised mean difference (SMD)}

\item{smd_digits}{(numeric) Number of decimal places to express SMD}

\item{bold_var}{(logical) If TRUE, variable sub-headers will be wrapped with **}
}
\value{
(data.frame) A tableby table appended with marker and SMD
}
\description{
\code{extend_tableby} is a wrapper around \link[arsenal:tableby]{arsenal::tableby} that adds the
following to a tableby table:
\enumerate{
\item Marker for enabling the joining of two tableby tables together
\item Standardised mean difference (SMD) between the groups
}
}
\examples{
library(arsenal)
library(labelled)
library(MatchIt)
data(lalonde)

# Add labels (optional) -----------------------------------------------------
var_label(lalonde) <- list(
  age = "Age",
  educ = "Education",
  race = "Race",
  married = "Marital status"
)

# Create tableby table ------------------------------------------------------
tableby(treat ~ age + educ + race + married,
  data = lalonde,
  control = tableby.control(numeric.stats = c("Nmiss", "meansd", "medianq1q3", "range"))
) \%>\%
  extend_tableby(lalonde)

}
